requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiofiles==23.2.1
aiosqlite==0.19.0
asyncio-throttle==1.0.2
colorlog==6.7.0
humanize==4.8.0
